gen4_gvm_gy: Create dtb.img with DT table structure
The dtb.img created as part of the Android build system
is a flattened list of device trees. This is an issue
for VM loading as RM VM relies on a device tree header
to differentiate between the different device trees.

Let's add support to build the device tree header and
package the generated dtb.img, with the header, as part
of the VM image. This allows the RM VM device tree selection
logic to select the expected device tree.

Change-Id: Iee5e2d70f4ce10531250f12bd7a3401043acefd7
Signed-off-by: Wei Liu <quic_wliu8@quicinc.com>